Refactor Lambda stuff out
