+++ crates/uplc/src/program_builder/tests.rs
@@ -3,16 +3,24 @@ use crate::parser;
 use crate::program_builder::constant::WithConstant;
 use proptest::prelude::*;
 
-proptest! {
-    #[test]
-    fn build_named__with_version(
+prop_compose! {
+    fn arb_version()(
         maj: isize,
         min: isize,
         patch: isize,
-    ) {
+    ) -> (usize, usize, usize){
         let maj = maj.abs() as usize;
         let min = min.abs() as usize;
         let patch = patch.abs() as usize;
+        (maj, min, patch)
+    }
+}
+
+proptest! {
+    #[test]
+    fn build_named__with_version(
+        (maj, min, patch) in arb_version(),
+    ) {
         let code = format!(r"(program
                            {}.{}.{}
                            (con integer 11)
@@ -21,33 +29,4 @@ proptest! {
         let actual = Builder::new(maj, min, patch).with_constant_int(11).build_named();
         assert_eq!(expected, actual);
     }
-}
-
-#[test]
-fn build_named__with_lam() {
-    let code = r"(program
-                       1.2.3
-                       (lam i_0 (con integer 1))
-                     )";
-    let expected = parser::program(code).unwrap();
-    let actual = Builder::new(1, 2, 3)
-        .with_lambda("i_0")
-        .with_constant_int(1)
-        .build_named();
-    assert_eq!(expected, actual);
-}
-
-#[test]
-fn build_named__with_nested_lam() {
-    let code = r"(program
-                       1.2.3
-                       (lam i_0 (lam i_1 (con integer 1)))
-                     )";
-    let expected = parser::program(code).unwrap();
-    let actual = Builder::new(1, 2, 3)
-        .with_lambda("i_0")
-        .with_lambda("i_1")
-        .with_constant_int(1)
-        .build_named();
-    assert_eq!(expected, actual);
 }
